Various fixes

+++ submodules/LegacyMediaPickerUI/Sources/LegacyPaintStickersContext.swift
@@ -461,7 +461,7 @@ public final class LegacyPaintEntityRenderer: NSObject, TGPhotoPaintEntityRender
             }
         }
         
-        func gcd(_ a: Int32, _ b: Int32) -> Int32 {
+        func gcd(_ a: Int64, _ b: Int64) -> Int64 {
             let remainder = a % b
             if remainder != 0 {
                 return gcd(b, remainder)
@@ -470,7 +470,7 @@ public final class LegacyPaintEntityRenderer: NSObject, TGPhotoPaintEntityRender
             }
         }
         
-        func lcm(_ x: Int32, _ y: Int32) -> Int32 {
+        func lcm(_ x: Int64, _ y: Int64) -> Int64 {
             return x / gcd(x, y) * y
         }
         
@@ -480,7 +480,7 @@ public final class LegacyPaintEntityRenderer: NSObject, TGPhotoPaintEntityRender
             let minDuration: Double = 3.0
             if durations.count > 1 {
                 let reduced = durations.reduce(1.0) { lhs, rhs -> Double in
-                    return Double(lcm(Int32(lhs * 10.0), Int32(rhs * 10.0)))
+                    return Double(lcm(Int64(lhs * 10.0), Int64(rhs * 10.0)))
                 }
                 result = min(6.0, Double(reduced) / 10.0)
             } else if let duration = durations.first {
